backend/requirements.txt
Flask==3.0.0
PyJWT[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0
